requirements.txt
customtkinter==5.2.0
flatbuffers==24.3.25
numpy==1.24.3
onnx==1.15.0
onnx2tf==1.17.5
onnx_simplifier==0.4.36
onnxruntime==1.17.1
onnxsim==0.4.36
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pip==24.2
pyserial==3.5
pytorch_lightning==2.0.0
pyts==0.13.0
PyYAML==6.0.1
setuptools==72.1.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.65.2
